one more findbugs warning

+++ java/src/jmri/jmrit/withrottle/MultiThrottleController.java
@@ -97,7 +97,7 @@ public class MultiThrottleController extends ThrottleController {
                }
             } else {
                if( Math.abs(lastSentSpeed.peek().floatValue()-currentSpeed)<0.0005 ) {
-                  Float f = (Float)lastSentSpeed.poll(); // remove the value from the list.
+                  Float f = lastSentSpeed.poll(); // remove the value from the list.
                   log.debug("removed value {} from queue",f);
                }
             }
